Move scaling's RAL handler to shardingsphere-scaling-distsql-handler module (#11180)
