+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/java/org/apache/shardingsphere/proxy/backend/text/distsql/ral/update/UpdatableRALBackendHandlerFactory.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.proxy.backend.text.distsql.ral;
+package org.apache.shardingsphere.proxy.backend.text.distsql.ral.update;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
