+++ shardingsphere-scaling/shardingsphere-scaling-distsql/shardingsphere-scaling-distsql-handler/src/main/java/org/apache/shardingsphere/scaling/distsql/exception/ScalingJobOperateException.java
@@ -15,19 +15,18 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.proxy.backend.exception;
+package org.apache.shardingsphere.scaling.distsql.exception;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
+import org.apache.shardingsphere.infra.exception.ShardingSphereException;
 
 /**
  * Scaling job operate exception.
  */
-@RequiredArgsConstructor
-@Getter
-public final class ScalingJobOperateException extends BackendException {
+public final class ScalingJobOperateException extends ShardingSphereException {
     
     private static final long serialVersionUID = 7598088400647370901L;
     
-    private final String message;
+    public ScalingJobOperateException(final String reason) {
+        super(reason);
+    }
 }
