+++ shardingsphere-scaling/shardingsphere-scaling-distsql/shardingsphere-scaling-distsql-handler/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.distsql.update.RALUpdater
@@ -15,4 +15,7 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.proxy.backend.text.distsql.rql.DataSourceQueryResultSet
+org.apache.shardingsphere.scaling.distsql.handler.StartScalingJobUpdater
+org.apache.shardingsphere.scaling.distsql.handler.StopScalingJobUpdater
+org.apache.shardingsphere.scaling.distsql.handler.ResetScalingJobUpdater
+org.apache.shardingsphere.scaling.distsql.handler.DropScalingJobUpdater
